backend/events/pyproject.toml
[tool.poetry]
name = "events service"
version = "0.1.0"
description = ""
authors = ["ligets <legen2a208@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
alembic = "^1.13.3"
asyncpg = "^0.29.0"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
python-dotenv = "^1.0.1"
datetime = "^5.5"
pyjwt = "^2.9.0"
aio-pika = "^9.4.3"
bs4 = "^0.0.2"
python-multipart = "^0.0.19"
redis = "^5.2.1"
fastapi-cache2 = "^0.2.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"

[tool.pytest.ini_options]
pythonpath = [
    ".", "src"
]
asyncio_mode="auto"
asyncio_default_fixture_loop_scope="session"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
